Move workflow to lib/worker/run-workflow
